masthead improvements, footer content
added in some more styling elements for the header. particulalry the
movbile version and headers for dropdowns on mobile.

Filled out the footer content with social data and legals. Also styled
out the social iframes.

TODO: will not to seperate the elements within the app.scss file as its
starting to get a little cluttered.